+++ frameworks/react/components/display/table-row/TableRow.d.ts
@@ -4,6 +4,8 @@ export interface TableRowProps {
 
   children?: React.ReactNode;
 
+  disabled?: boolean;
+
   onClick?: () => void;
 }